break down the submission flow
